[FIX] "runtime" generated ros files in psa
when btl files are present, but no .ros files, SBE created the ros
files.
SBE runs after the templates are generated, so the ros files are not in
the template.
This was solved by adding the stems of the blt files to the list of ros
files.
Possible issue: not unique files?